requirements.txt
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
google-api-core==2.6.0
google-auth==2.6.0
google-cloud-texttospeech==2.10.1
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
pdfminer.six==20211012
pdfplumber==0.6.0
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Wand==0.6.7
